@@ -55,10 +55,10 @@
         "phpunit/phpunit": "^9.5.13",
         "psalm/plugin-phpunit": "^0.16",
         "psalm/plugin-symfony": "^3.0",
-        "rector/rector": "^0.14",
+        "rector/rector": "^0.15",
         "symfony/browser-kit": "^5.4 || ^6.0",
         "symfony/error-handler": "^5.4 || ^6.0",
-        "symfony/phpunit-bridge": "^6.0",
+        "symfony/phpunit-bridge": "^6.1",
         "symfony/stopwatch": "^5.4 || ^6.0",
         "vimeo/psalm": "^4.1"
     },
