@@ -88,7 +88,8 @@ final class StrategyManager implements StrategyManagerInterface
 
             // Convert throwable to exception
             if (!$exception instanceof \Exception) {
-                $exception = new \Exception($exception->getMessage(), (int) $exception->getCode(), $exception);
+                /** @psalm-suppress InvalidScalarArgument */
+                $exception = new \Exception($exception->getMessage(), $exception->getCode(), $exception);
             }
 
             $response = $renderer->render($exception, $block, $response);
